+++ src/main/java/com/sfjs/gql/resolvers/Signup.java
@@ -45,16 +45,4 @@ public class Signup {
     return result;
   }
 
-  @MutationMapping(name = "resetPassword")
-  public Result resetPassword(@Argument(name = "email") String email,
-      @Argument(name = "password") String password,
-      @Argument(name = "token") String token) {
-    return signupService.resetPassword(email, password, token);
-  }
-
-  @MutationMapping(name = "generateResetPasswordToken")
-  public Result generateResetPasswordToken(
-      @Argument(name = "email") String email) {
-    return signupService.generateResetPasswordToken(email);
-  }
 }
